Make separate script for parsing functions
