+++ ins.sh
@@ -1,0 +1,2 @@
+pip uninstall simc -y
+python setup.py install
