+++ test.c
@@ -1,0 +1,10 @@
+#include <stdio.h>
+
+int main() {
+	char* i;
+	printf("Enter a string: ");
+	scanf("%s", i);
+	printf("%s", i);
+
+	return 0;
+}
